Replace PDL_Index by PDL_Indx to fix Types.pm.PL
NOTE: The choice of the typename is a result of the current hack
for generating the various type conversion routines from some of
the fields in Types.pm.PL structures.  I hope we can clean this
up for PDL3 to avoid the confusing mess.  Another goal for PDL3
might be to collect all the type generation and use code into a
real PDL core.  Right now it is spread amongst PDL/Basic/Core,
PDL/Basic/Gen and PDL/Basic/Slices (at least!).